+++ src/main/java/com/backend/service/IAddressService.java
@@ -24,7 +24,4 @@ public interface IAddressService {
 
     ServiceResult<Address> deleteAddress(AddressRequest addressRequest);
 
-    ServiceResult<List<AccountAddress>> getAllAddressAndAccount(Long id);
-
-
 }
